+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Deep compare stack traces when formatting.
+
 ## [0.7.1] - 2023-09-20
 
 ### Fixed