+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.0] - 2023-10-12
+
 ### Added
 
 - Detect stack traces in errors made by `github.com/rotisserie/eris` package.
@@ -132,7 +134,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First public release.
 
-[unreleased]: https://gitlab.com/tozd/go/errors/-/compare/v0.7.2...main
+[unreleased]: https://gitlab.com/tozd/go/errors/-/compare/v0.8.0...main
+[0.8.0]: https://gitlab.com/tozd/go/errors/-/compare/v0.7.2...v0.8.0
 [0.7.2]: https://gitlab.com/tozd/go/errors/-/compare/v0.7.1...v0.7.2
 [0.7.1]: https://gitlab.com/tozd/go/errors/-/compare/v0.7.0...v0.7.1
 [0.7.0]: https://gitlab.com/tozd/go/errors/-/compare/v0.6.0...v0.7.0